fix: fix file locations
