+++ src/modules/testing-tree/index.ts
@@ -1,3 +1,4 @@
-export { TestingTree } from "./model/testing-tree";
+export { TestingTree } from "./store/testing-tree";
+export {TestingTreeSingleton} from "./store/testing-tree-singleton"
 export { GroupNode } from "./classes/group-node";
 export { type TestNode } from "./types";
