+++ src/modules/tests/classes/test-group.ts
@@ -1,4 +1,4 @@
-import { TestInfo, TestStatus } from "../types";
+import { TestInfo, TestStatus } from "../types/core";
 
 export class TestsGroup implements TestInfo {
     public status = TestStatus.NOT_RUNED;
