+++ src/modules/tests/index.ts
@@ -2,5 +2,5 @@ export { SampleTest } from "./classes/sample-test";
 export { TestsGroup } from "./classes/test-group";
 export { BaseTestFactory } from "./factories/base-test-factory";
 export { TestFactory } from "./factories/test-factory";
-export { TestStatus } from "./types";
-export type { TestInfo, AvailableTestStates } from "./types";
+export { TestStatus } from "./types/core";
+export type { TestInfo, AvailableTestStates, TestResult } from "./types/core";
