--- src/shared/types/tests.ts
@@ -1,5 +1,0 @@
-export type TestResult = {
-    title: string;
-    status: "succeed" | "failed" | "error" | "not runed";
-    children?: TestResult[];
-};
